Restructure helper functions into a new class
