+++ util/src/main/java/tech/pegasys/artemis/util/hashToG2/FP2Immutable.java
@@ -13,7 +13,7 @@
 
 package tech.pegasys.artemis.util.hashToG2;
 
-import static tech.pegasys.artemis.util.hashToG2.Util.P;
+import static tech.pegasys.artemis.util.hashToG2.Helper.P;
 import static tech.pegasys.artemis.util.hashToG2.Util.bigFromHex;
 
 import java.util.Objects;
@@ -25,54 +25,14 @@ import org.apache.milagro.amcl.BLS381.FP2;
  * It's safer and simpler to wrap FP2 objects into an immutable type. The underlying FP2 type is
  * highly mutable. As a benefit, this class allows us to chain field operations nicely - this makes
  * dealing with Milagro much less tiresome.
+ *
+ * <p>However, this does involve many more memory copies. For now, the coding convenience outweighs
+ * any performance concerns. We can optimise later if it is a problem.
  */
-public class FP2Immutable {
+public final class FP2Immutable {
 
   static final FP2Immutable ZERO = new FP2Immutable(0);
   static final FP2Immutable ONE = new FP2Immutable(1);
-
-  /**
-   * Raise a field point to an integer exponent.
-   *
-   * @param x the field point
-   * @param exponent the exponent
-   * @return the field point raised to the exponent and reduced
-   */
-  static FP2Immutable pow(FP2Immutable x, int exponent) {
-    if (exponent == 0) return ONE;
-    if (exponent == 1) return x;
-    if (exponent == 2) return x.sqr().reduce();
-    FP2Immutable result = ONE;
-    while (exponent > 0) {
-      if (exponent % 2 == 1) {
-        result = result.mul(x);
-      }
-      x = x.sqr();
-      exponent /= 2;
-    }
-    return result.reduce();
-  }
-
-  /**
-   * Raise a field point to a DBIG exponent.
-   *
-   * @param x the field point
-   * @param exponent the DBIG exponent
-   * @return the field point raised to the DBIG exponent and reduced
-   */
-  static FP2Immutable pow(FP2Immutable x, DBIG exponent) {
-    FP2Immutable result = ONE;
-    DBIGExtended exp = new DBIGExtended(exponent);
-    exp.norm();
-    while (!exp.iszilch()) {
-      if (exp.parity() == 1) {
-        result = result.mul(x);
-      }
-      x = x.sqr();
-      exp.shr(1);
-    }
-    return result.reduce();
-  }
 
   private final FP2 fp2;
 
@@ -125,54 +85,67 @@ public class FP2Immutable {
     this.fp2 = new FP2(bigFromHex(hex1), bigFromHex(hex2));
   }
 
+  /** Wrap FP2 sqr() method to return a result */
   FP2Immutable sqr() {
     FP2 result = new FP2(this.fp2);
     result.sqr();
     return new FP2Immutable(result);
   }
 
+  /** Wrap FP2 mul() method to return a result */
   FP2Immutable mul(FP2Immutable a) {
     FP2 result = new FP2(this.fp2);
     result.mul(a.fp2);
     return new FP2Immutable(result);
   }
 
+  /** Wrap FP2 add() method to return a result */
   FP2Immutable add(FP2Immutable a) {
     FP2 result = new FP2(this.fp2);
     result.add(a.fp2);
     return new FP2Immutable(result);
   }
 
+  /** Wrap FP2 dbl() method to return a result */
   FP2Immutable dbl() {
     FP2 result = new FP2(this.fp2);
     result.add(this.fp2);
     return new FP2Immutable(result);
   }
 
+  /** Wrap FP2 sub() method to return a result */
   FP2Immutable sub(FP2Immutable a) {
     FP2 result = new FP2(this.fp2);
     result.sub(a.fp2);
     return new FP2Immutable(result);
   }
 
+  /** Wrap FP2 neg() method to return a result */
   FP2Immutable neg() {
     FP2 result = new FP2(this.fp2);
     result.neg();
     return new FP2Immutable(result);
   }
 
+  /** Wrap FP2 reduce() method to return a result */
   FP2Immutable reduce() {
     FP2 result = new FP2(this.fp2);
     result.reduce();
     return new FP2Immutable(result);
   }
 
+  /** Wrap FP2 inverse() method to return a result */
   FP2Immutable inverse() {
     FP2 result = new FP2(this.fp2);
     result.inverse();
     return new FP2Immutable(result);
   }
 
+  /**
+   * Test whether the element is zero.
+   *
+   * @return true if the element is zero, false otherwise
+   */
   boolean iszilch() {
     return fp2.iszilch();
   }
@@ -192,6 +165,49 @@ public class FP2Immutable {
       cmp = BIG.comp(fp2.getA(), thresh);
     }
     return cmp == 0 ? 1 : -cmp;
+  }
+
+  /**
+   * Raise this element to an integer exponent.
+   *
+   * @param exponent the exponent
+   * @return the field point raised to the exponent and reduced
+   */
+  FP2Immutable pow(int exponent) {
+    if (exponent == 0) return ONE;
+    if (exponent == 1) return this;
+    if (exponent == 2) return this.sqr().reduce();
+    FP2Immutable result = ONE;
+    FP2Immutable tmp = new FP2Immutable(this);
+    while (exponent > 0) {
+      if (exponent % 2 == 1) {
+        result = result.mul(tmp);
+      }
+      tmp = tmp.sqr();
+      exponent /= 2;
+    }
+    return result.reduce();
+  }
+
+  /**
+   * Raise this element to a DBIG exponent.
+   *
+   * @param exponent the DBIG exponent
+   * @return the field point raised to the DBIG exponent and reduced
+   */
+  FP2Immutable pow(DBIG exponent) {
+    FP2Immutable result = ONE;
+    DBIGExtended exp = new DBIGExtended(exponent);
+    exp.norm();
+    FP2Immutable tmp = new FP2Immutable(this);
+    while (!exp.iszilch()) {
+      if (exp.parity() == 1) {
+        result = result.mul(tmp);
+      }
+      tmp = tmp.sqr();
+      exp.shr(1);
+    }
+    return result.reduce();
   }
 
   FP2 getFp2() {
