+++ util/src/test/java/tech/pegasys/artemis/util/hashToG2/HelperTest.java
@@ -17,21 +17,19 @@ import static java.nio.charset.StandardCharsets.UTF_8;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 import static tech.pegasys.artemis.util.hashToG2.FP2Immutable.ONE;
-import static tech.pegasys.artemis.util.hashToG2.Util.HKDF_Expand;
-import static tech.pegasys.artemis.util.hashToG2.Util.HKDF_Extract;
-import static tech.pegasys.artemis.util.hashToG2.Util.HMAC_SHA256;
-import static tech.pegasys.artemis.util.hashToG2.Util.ROOTS_OF_UNITY;
+import static tech.pegasys.artemis.util.hashToG2.Helper.HKDF_Expand;
+import static tech.pegasys.artemis.util.hashToG2.Helper.HKDF_Extract;
+import static tech.pegasys.artemis.util.hashToG2.Helper.HMAC_SHA256;
+import static tech.pegasys.artemis.util.hashToG2.Helper.ROOTS_OF_UNITY;
+import static tech.pegasys.artemis.util.hashToG2.Helper.clear_h2;
+import static tech.pegasys.artemis.util.hashToG2.Helper.hashToBase;
+import static tech.pegasys.artemis.util.hashToG2.Helper.iso3;
+import static tech.pegasys.artemis.util.hashToG2.Helper.mapToCurve;
+import static tech.pegasys.artemis.util.hashToG2.Helper.mx_chain;
+import static tech.pegasys.artemis.util.hashToG2.Helper.onCurveG2;
+import static tech.pegasys.artemis.util.hashToG2.Helper.os2ip;
+import static tech.pegasys.artemis.util.hashToG2.Helper.psi;
 import static tech.pegasys.artemis.util.hashToG2.Util.bigFromHex;
-import static tech.pegasys.artemis.util.hashToG2.Util.clear_h2;
-import static tech.pegasys.artemis.util.hashToG2.Util.hashToBase;
-import static tech.pegasys.artemis.util.hashToG2.Util.iso3;
-import static tech.pegasys.artemis.util.hashToG2.Util.mapToCurve;
-import static tech.pegasys.artemis.util.hashToG2.Util.mx_chain;
-import static tech.pegasys.artemis.util.hashToG2.Util.onCurveG2;
-import static tech.pegasys.artemis.util.hashToG2.Util.os2ip;
-import static tech.pegasys.artemis.util.hashToG2.Util.psi;
-import static tech.pegasys.artemis.util.hashToG2.Util.xi_2;
-import static tech.pegasys.artemis.util.hashToG2.Util.xi_2Pow2;
 
 import java.math.BigInteger;
 import org.apache.milagro.amcl.BLS381.BIG;
@@ -42,7 +40,7 @@ import org.apache.tuweni.bytes.Bytes;
 import org.junit.jupiter.api.Test;
 import tech.pegasys.artemis.util.mikuli.G2Point;
 
-class UtilTest {
+class HelperTest {
 
   /*
    * HMAC_SHA256 Tests
@@ -287,11 +285,6 @@ class UtilTest {
     }
   }
 
-  @Test
-  void xi2sqEqualsXi2SquaredTest() {
-    assertEquals(xi_2Pow2, xi_2.sqr());
-  }
-
   /*
    * The remaining tests use data generated by the reference implementation at
    * https://github.com/kwantam/bls_sigs_ref/tree/master/python-impl
@@ -366,7 +359,6 @@ class UtilTest {
     assertEquals(expected, mx_chain(a));
   }
 
-  // Unit tests based on tracing the Python code with standard reference data
   @Test
   void hashToBaseTest0() {
     FP2Immutable expected =
